Update histogram_fit.py
Overhauled histogram_fit.py. This entails more customization and an improved fitting function.
Maximum offset can be set manually, as can the number of bins. The y range changes so the graph fills the majority of the image.
The reason the bin number is not fixed is because if I set it too low with data with an outlier, all the offsets will end up in one bin.
The curve is now using scipy.stats.lognorm. Lognorm.fit is used to fit the curve to the unbinned data.